Fixed type error

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,8 +8,8 @@
   */
 int main(void)
 {
-	int i = 0;
-	int j = 1;
+	long i = 0;
+	long j = 1;
 	int count;
 	long next;
